Create BinarySearchExample.cs
Best for SORTED ARRAY.

Compare x with the middle element.
If x matches with the middle element, we return the mid index.
Else If x is greater than the mid element, then x can only lie in right half subarray after the mid element. So we recur for the right half.
Else (x is smaller) recur for the left half.