Create ProductofTwoConsecutiveIntegersCodility.cs
Write a function solution that given two integers A and B, returns the number of integers from the range [A..B](ends are included) which can be expressed as the product of two consecutive integers, that is X*(X+1), for some integer X.


Examples:

Given A =6 and B=20, the function should return 3. These integers are 6=2*3, 12=3*4 and 20=4*5.

Given A=21 and B=29, the function should return 0. There are no integers of the form X*(X+1) in this interval.

Write an efficient algorithm for the following assumptions:

A and B are integers within the range [1..1,000,000,000]

A<=B